后端开发与修复

+++ demo01/src/com/test/pojo/Cases.java
@@ -3,20 +3,20 @@ package com.test.pojo;
 
 public class Cases {
 
-  private long id;
+  private Integer id;
   private String title;
   private String content;
-  private long commentId;
+  private Integer commentId;
   private String imgpath;
   private String grouptype;
   private String date;
 
 
-  public long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
@@ -38,11 +38,11 @@ public class Cases {
     this.content = content;
   }
 
-  public long getCommentId() {
+  public Integer getCommentId() {
     return commentId;
   }
 
-  public void setCommentId(long commentId) {
+  public void setCommentId(Integer commentId) {
     this.commentId = commentId;
   }
